rename fsdp.py to fully_sharded_data_parallel.py
