+++ oslo/torch/nn/parallel/distributed/__init__.py
@@ -1,4 +1,4 @@
-from oslo.torch.nn.parallel.distributed.tensor_parallel import *
+from oslo.torch.nn.parallel.distributed.data_parallel import *
 from oslo.torch.nn.parallel.distributed.tensor_parallel import *
 from oslo.torch.nn.parallel.distributed.pipeline_parallel import *
 from oslo.torch.nn.parallel.distributed.wrap import (
